requirements.txt
asgiref==3.6.0
Django==4.0.6
django-filter==23.1
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.3
mysqlclient==2.1.1
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
